LayoutTests/media/media-source/media-source-samples-resolution-change-expected.txt


EXPECTED (source.readyState == 'closed') OK
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer('video/mp4; codecs="avc1.4d401f"'))
EVENT(updateend)
Appended 6 seconds of 320x240
EVENT(updateend)
Replaced first 3 seconds with 480x360
{PTS(0.000), DTS(0.000), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.033), DTS(0.033), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.067), DTS(0.067), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.100), DTS(0.100), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.133), DTS(0.133), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.167), DTS(0.167), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.200), DTS(0.200), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.233), DTS(0.233), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.267), DTS(0.267), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.300), DTS(0.300), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.333), DTS(0.333), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.367), DTS(0.367), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.400), DTS(0.400), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.433), DTS(0.433), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.467), DTS(0.467), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.500), DTS(0.500), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.533), DTS(0.533), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.567), DTS(0.567), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.600), DTS(0.600), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.633), DTS(0.633), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.667), DTS(0.667), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.700), DTS(0.700), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.733), DTS(0.733), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.767), DTS(0.767), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.800), DTS(0.800), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.833), DTS(0.833), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.867), DTS(0.867), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.900), DTS(0.900), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.933), DTS(0.933), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(0.967), DTS(0.967), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.000), DTS(1.000), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.033), DTS(1.033), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.067), DTS(1.067), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.100), DTS(1.100), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.133), DTS(1.133), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.167), DTS(1.167), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.200), DTS(1.200), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.233), DTS(1.233), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.267), DTS(1.267), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.300), DTS(1.300), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.333), DTS(1.333), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.367), DTS(1.367), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.400), DTS(1.400), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.433), DTS(1.433), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.467), DTS(1.467), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.500), DTS(1.500), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.533), DTS(1.533), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.567), DTS(1.567), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.600), DTS(1.600), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.633), DTS(1.633), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.667), DTS(1.667), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.700), DTS(1.700), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.733), DTS(1.733), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.767), DTS(1.767), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.800), DTS(1.800), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.833), DTS(1.833), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.867), DTS(1.867), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.900), DTS(1.900), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.933), DTS(1.933), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(1.967), DTS(1.967), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.000), DTS(2.000), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.033), DTS(2.033), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.067), DTS(2.067), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.100), DTS(2.100), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.133), DTS(2.133), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.167), DTS(2.167), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.200), DTS(2.200), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.233), DTS(2.233), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.267), DTS(2.267), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.300), DTS(2.300), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.333), DTS(2.333), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.367), DTS(2.367), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.400), DTS(2.400), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.433), DTS(2.433), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.467), DTS(2.467), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.500), DTS(2.500), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.533), DTS(2.533), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.567), DTS(2.567), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.600), DTS(2.600), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.633), DTS(2.633), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.667), DTS(2.667), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.700), DTS(2.700), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.733), DTS(2.733), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.767), DTS(2.767), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.800), DTS(2.800), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.833), DTS(2.833), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.867), DTS(2.867), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.900), DTS(2.900), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.933), DTS(2.933), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(2.967), DTS(2.967), duration(0.033), presentationSize(480.000000x360.000000)}
{PTS(3.000), DTS(3.000), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.033), DTS(3.033), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.067), DTS(3.067), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.100), DTS(3.100), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.133), DTS(3.133), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.167), DTS(3.167), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.200), DTS(3.200), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.233), DTS(3.233), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.267), DTS(3.267), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.300), DTS(3.300), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.333), DTS(3.333), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.367), DTS(3.367), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.400), DTS(3.400), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.433), DTS(3.433), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.467), DTS(3.467), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.500), DTS(3.500), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.533), DTS(3.533), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.567), DTS(3.567), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.600), DTS(3.600), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.633), DTS(3.633), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.667), DTS(3.667), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.700), DTS(3.700), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.733), DTS(3.733), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.767), DTS(3.767), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.800), DTS(3.800), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.833), DTS(3.833), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.867), DTS(3.867), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.900), DTS(3.900), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.933), DTS(3.933), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(3.967), DTS(3.967), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.000), DTS(4.000), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.033), DTS(4.033), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.067), DTS(4.067), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.100), DTS(4.100), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.133), DTS(4.133), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.167), DTS(4.167), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.200), DTS(4.200), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.233), DTS(4.233), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.267), DTS(4.267), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.300), DTS(4.300), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.333), DTS(4.333), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.367), DTS(4.367), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.400), DTS(4.400), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.433), DTS(4.433), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.467), DTS(4.467), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.500), DTS(4.500), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.533), DTS(4.533), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.567), DTS(4.567), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.600), DTS(4.600), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.633), DTS(4.633), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.667), DTS(4.667), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.700), DTS(4.700), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.733), DTS(4.733), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.767), DTS(4.767), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.800), DTS(4.800), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.833), DTS(4.833), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.867), DTS(4.867), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.900), DTS(4.900), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.933), DTS(4.933), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(4.967), DTS(4.967), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.000), DTS(5.000), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.033), DTS(5.033), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.067), DTS(5.067), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.100), DTS(5.100), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.133), DTS(5.133), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.167), DTS(5.167), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.200), DTS(5.200), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.233), DTS(5.233), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.267), DTS(5.267), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.300), DTS(5.300), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.333), DTS(5.333), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.367), DTS(5.367), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.400), DTS(5.400), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.433), DTS(5.433), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.467), DTS(5.467), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.500), DTS(5.500), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.533), DTS(5.533), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.567), DTS(5.567), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.600), DTS(5.600), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.633), DTS(5.633), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.667), DTS(5.667), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.700), DTS(5.700), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.733), DTS(5.733), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.767), DTS(5.767), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.800), DTS(5.800), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.833), DTS(5.833), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.867), DTS(5.867), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.900), DTS(5.900), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.933), DTS(5.933), duration(0.033), presentationSize(320.000000x240.000000)}
{PTS(5.967), DTS(5.967), duration(0.033), presentationSize(320.000000x240.000000)}
END OF TEST
